requirements.txt
# Core dependencies
functions-framework==3.5.0
google-cloud-secret-manager==2.18.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# API clients
tavily-python==0.3.3
openai==1.10.0
langchain==0.1.6
langchain-openai==0.0.5

# Utilities
tenacity==8.2.3
requests==2.31.0
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Type checking (optional, for development)
mypy==1.8.0
types-requests==2.31.0.20240125